fix: fix bug in cloudinary utility

+++ src/utiles/cloudinary.js
@@ -11,11 +11,13 @@ cloudinary.config({
 const uploadOnCloudinary = async(localFilePath)=>{
     try {
         if (!localFilePath) return null
-        const response = await cloudinary.uploader.upload(localFilePath, {resource_type:auto})
+        const response = await cloudinary.uploader.upload(localFilePath, {resource_type:'auto'})
         fs.unlinkSync(localFilePath)
         return response
     } catch (error) {
-        fs.unlinkSync(localFilePath)
+        if (fs.existsSync(localFilePath)) {
+          fs.unlinkSync(localFilePath);
+        }
         return null
     }
 }
